+++ src/main/java/com/wandoujia/hackday/apppath/model/PointModel.java
@@ -9,8 +9,8 @@ public class PointModel {
     private String data;
     private Date start;
     private Long duration;
-    private Long longitude;
-    private Long latitude;
+    private Double longitude;
+    private Double latitude;
 
     public Long getUid() {
         return uid;
@@ -60,19 +60,19 @@ public class PointModel {
         this.data = data;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 }
